todo: fix sharp errors (might be environ problem)

Conflicts:
	README.md
	package.json
	src/commands/games/blackjack.ts
	src/commands/games/liarsdice.ts
	src/commands/index.ts
	src/index.ts
	src/types/index.ts
	yarn.lock

+++ src/commands/messages/cowsay.ts
@@ -1,17 +1,21 @@
-import { say } from 'cowsay';
+import { say } from '@wiisportsresorts/cowsay';
+import * as cows from '@wiisportsresorts/cowsay/lib/cows';
 import { Message } from 'discord.js';
 import yargsParser from 'yargs-parser';
 
 import { Command } from '..';
-import { CmdArgs } from '../../types';
-import { Embed } from '../../util';
+import { Context } from '../../types';
+import { codeBlock, Embed } from '../../util';
 
 export class CommandCowsay implements Command {
   cmd = 'cowsay';
-  yargsSchema: yargsParser.Options = {
-    boolean: ['delete'],
+  yargs: yargsParser.Options = {
+    boolean: ['delete', 'list'],
+    string: ['cow'],
     alias: {
       delete: 'd',
+      cow: 'f',
+      list: 'l',
     },
     default: {
       delete: false,
@@ -21,19 +25,32 @@ export class CommandCowsay implements Command {
     usage: 'cowsay [-d, --delete] <...msg>',
     description: 'you know what it does (`--delete` deletes source command)',
   };
-  async executor(cmdArgs: CmdArgs): Promise<void | Message> {
-    const { msg, args } = cmdArgs;
+  async execute(context: Context): Promise<void | Message> {
+    const { msg, args } = context;
+
+    if (args.list)
+      return msg.channel.send(Embed.info('Cows', codeBlock(Object.keys(cows).join(', '))));
 
     if (args._.length == 0 || /^\s+$/.test(args._.join(' ')))
-      return msg.channel.send(Embed.error('**nothing to say**'));
+      return msg.channel.send(Embed.error('nothing to say'));
 
     args.delete && msg.deletable && msg.delete();
 
-    return msg.channel.send(
-      `\`\`\`\n${say({
-        text: args._.join(' '),
-        W: 48,
-      })}\n\`\`\``
-    );
+    const cow = Object.keys(cows).includes(args.cow)
+      ? cows[args.cow as keyof typeof import('@wiisportsresorts/cowsay/lib/cows')]
+      : undefined;
+
+    if (args.cow && !cow) return msg.channel.send(Embed.error('Unknown cow'));
+
+    const text = say(args._.join(' '), {
+      W: 48,
+      cow,
+    }).replace(/```/g, "'''"); // prevent codeblock escaping
+
+    const messages = text
+      .match(/(.|\n){1,1990}\n/g)
+      ?.map(message => codeBlock(message)) as string[];
+
+    for (const text of messages) await msg.channel.send(text);
   }
 }
